requirements.txt
numpy>=1.26,<3
pandas>=2.2,<3
matplotlib>=3.8,<4
seaborn>=0.13,<1
scikit-learn>=1.6,<2
jupyter>=1.0,<2
ipykernel>=6.29,<7
nbconvert>=7.16,<8
nbformat>=5.10,<6
pyyaml>=6,<7
